pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docwen"
dynamic = ["version"]
description = "公文图表格式转换软件 - 支持 Word/Markdown/Excel 互转"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "AGPL-3.0-or-later"}
authors = [
    {name = "ZhengYX", email = "zhengyx91@hotmail.com"}
]
keywords = ["document", "converter", "markdown", "docx", "excel", "office", "公文", "表格"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]

dependencies = [
    # 图形界面
    "ttkbootstrap>=1.18.0",
    "tkinterdnd2>=0.4.3",
    "Pillow>=11.3.0,<12.0.0",
    
    # 文档处理
    "python-docx>=1.2.0",
    "emoji>=2.0.0",
    "openpyxl>=3.1.5",
    "lxml>=6.0.2",
    "latex2mathml>=3.77.0",
    "PyYAML>=6.0.2",
    "tomlkit>=0.13.3",
    "pandas>=2.3.3",
    "olefile>=0.47",  # OLE文件格式解析，用于精确识别DOC/XLS
    
    # PDF/OFD处理
    "pymupdf>=1.26.6",
    "pymupdf4llm>=0.1.8",
    "pymupdf_layout>=1.27.1",
    "pdf2docx>=0.5.6",
    "easyofd>=0.5.6",  # OFD文件转PDF
    
    # 数值计算
    "numpy>=2.3.4,<3.0.0",
    
    # 图片转换
    "img2pdf>=0.6.1",
    "pillow-heif>=1.1.1",
    
    # OCR支持
    "rapidocr_onnxruntime>=1.4.0",
    
    # Windows COM 自动化
    "pywin32>=311; platform_system == 'Windows'",
    
    # IPC 进程间通信
    "watchdog>=6.0.0",
]

[project.optional-dependencies]
dev = [
    # 构建与打包
    "Cython==3.1.6",
    "pyinstaller==6.16.0",
    "pre-commit>=4.0.0",
    "ruff>=0.15.1",
    "pyright>=1.1.408",
]
lint = [
    "ruff>=0.15.1",
    "pyright>=1.1.408",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
docwen = "docwen.cli_run:main"
docwen-gui = "docwen.gui_run:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["docwen*"]

[tool.setuptools.dynamic]
version = {attr = "docwen.__version__"}

[tool.setuptools.package-data]
docwen = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --strict-markers -m \"not integration and not slow\""
markers = [
    "unit: 单元测试",
    "integration: 集成测试（需要真实文件）",
    "slow: 耗时测试",
    "windows_only: 仅 Windows 环境",
]

[tool.coverage.run]
branch = true
source = ["docwen"]
omit = [
    "*/tests/*",
    "*/docwen/gui/*",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 35

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
include = ["src"]
executionEnvironments = [
    { root = "src", extraPaths = ["src"] },
    { root = "src/docwen/gui", extraPaths = ["src"], reportAttributeAccessIssue = "warning" },
]
exclude = [
    "**/__pycache__",
    ".git",
    ".ruff_cache",
    ".pytest_cache",
    ".venv",
    "venv",
    "build",
    "dist",
    "assets",
    "templates",
    "configs",
    "tests",
]
stubPath = "typings"

reportMissingTypeStubs = "none"
reportMissingImports = "none"
reportMissingModuleSource = "none"
reportUnusedImport = "error"
reportUnusedVariable = "error"
reportUnusedExpression = "none"
reportInvalidStringEscapeSequence = "error"
reportPrivateUsage = "warning"
reportUntypedFunctionDecorator = "none"
reportUntypedClassDecorator = "none"

reportUnknownMemberType = "none"
reportUnknownVariableType = "none"
reportUnknownArgumentType = "none"
reportUnknownParameterType = "none"
reportUnknownLambdaType = "none"

reportWildcardImportFromLibrary = "none"

reportUnnecessaryIsInstance = "warning"
reportUnnecessaryCast = "warning"
reportUnnecessaryComparison = "warning"
reportUnnecessaryContains = "warning"

reportGeneralTypeIssues = "error"
reportArgumentType = "error"
reportAssignmentType = "error"
reportCallIssue = "error"
reportIndexIssue = "error"
reportOptionalMemberAccess = "error"
reportOptionalOperand = "error"
reportOptionalSubscript = "error"
reportOptionalIterable = "error"
reportOperatorIssue = "error"
reportAttributeAccessIssue = "error"
reportReturnType = "error"

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src"]
extend-exclude = ["assets", "templates", "configs"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "W",
    "UP",
    "B",
    "SIM",
    "C4",
    "PIE",
    "RUF",
    "T20",
]
ignore = [
    "E501",
    "RUF001",
    "RUF002",
    "RUF003",
]

[tool.ruff.lint.per-file-ignores]
"src/docwen/gui/**/*.py" = ["F403", "F405"]
"src/docwen/gui_run.py" = ["SIM115"]
"src/docwen/ipc/single_instance.py" = ["SIM115"]
"src/docwen/cli/**/*.py" = ["T201"]
"src/docwen/cli_run.py" = ["T201"]
"src/docwen/i18n/locale_validator.py" = ["T201"]
"src/docwen/config/safe_logger.py" = ["T201"]
"tools/**/*.py" = ["T201"]
"tests/**/*.py" = ["B011"]

[tool.ruff.lint.isort]
known-first-party = ["docwen"]
